requirements.txt
pbr>=0.5.21,<1.0

anyjson>=0.3.3
argparse
iso8601>=0.1.8
lxml>=2.3
oslo.config>=1.2.0
six>=1.4.1

kombu>=2.4.8

-f http://tarballs.openstack.org/oslo.messaging/oslo.messaging-1.3.0a2.tar.gz#egg=oslo.messaging-1.3.0a2
oslo.messaging>=1.3.0a2
